Fix: Modify how to use button
사용방법을 알기 쉽지 않다는 피드백을 수용하여 사용방법 버튼을 메인에서도 클릭할 수 있도록 배치
