+++ src/shared/apis.js
@@ -3,7 +3,6 @@ import { getCookie } from './Cookie';
 
 const api = axios.create({
   baseURL: process.env.REACT_APP_API_URL,
-  // withCredentials: true,
 });
 
 // Alter defaults after instance has been created
